+++ CardPayments/src/main/java/com/paypal/android/cardpayments/CardClient.kt
@@ -208,10 +208,4 @@ class CardClient internal constructor(
         }
         return result
     }
-
-    /**
-     * Call this method at the end of the card flow to clear out all observers and listeners
-     */
-    fun removeObservers() {
-    }
 }