+++ CHANGELOG.md
@@ -15,6 +15,7 @@
     * Remove `CardResult` type
     * Remove `CardClient.vault(Context, CardRequest)` method
     * Remove `CardClient.completeAuthChallenge(Intent, String)` method
+    * Remove `CardClient.removeObservers()` method
 
 ## 2.0.0-beta1 (2024-11-20)
 * Breaking Changes